pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "petey-web"
version = "0.1.0"
description = "Web interface for Petey PDF Extractor"
requires-python = ">=3.11"
dependencies = [
    "petey",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "pyyaml",
    "firebase-admin",
    "google-cloud-firestore",
    "httpx",
    "pikepdf<10",
    "ocrmypdf",
    "anthropic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "firebase-admin",
    "google-cloud-firestore",
]

[tool.setuptools.packages.find]
include = ["server*"]
